requirements.txt
torchvision==0.4.0a0+6b959ee
tensorflow==1.13.1
tqdm==4.35.0
scikit_video==1.1.11
colored_traceback==0.3.0
matplotlib==3.1.1
h5py==2.10.0
imageio==2.5.0
scikit_image==0.16.2
torch==1.2.0
numpy==1.17.0
moviepy==1.0.0
baseline==0.2.1
generator==0.1-alpha
ipdb==0.12.3
Pillow==7.0.0
skimage==0.0
tensorboardX==2.0
visdom==0.1.8.9
